cleanup: types
