+++ src/components/post/Post.tsx
@@ -1,9 +1,10 @@
 import { memo } from "react";
 import { Card } from "../card/Card";
 import { TagsBar } from "../tags-bar/TagsBar";
-import { formattedDate, PostData } from "../../helpers/PostData";
-import "./Post.css";
+import { formattedDate } from "../../helpers/PostData";
+import { PostData } from "../../types/PostData"
 import { PostContent } from "./content/PostContent";
+import "./Post.css";
 
 type Props = {
   post: PostData;
