+++ src/containers/projects/Projects.tsx
@@ -2,7 +2,7 @@ import { memo, useEffect, useState } from "react";
 import { PageContainer } from "../../components/page-container/PageContainer";
 import { fetchJsonData } from "../../helpers/DataReader.ts";
 import { Spinner } from "../../components/spinner/Spinner.tsx";
-import { ProjectData } from "../../helpers/ProjectData.ts";
+import { ProjectData } from "../../types/ProjectData.ts";
 import { ProjectCard } from "./project-card/ProjectCard.tsx";
 import { WipMessage } from "../../components/wip-message/WipMessage.tsx";
