+++ src/containers/projects/project-card/ProjectCard.tsx
@@ -1,6 +1,6 @@
 import { memo } from "react";
 import { Card } from "../../../components/card/Card.tsx";
-import { ProjectData } from "../../../helpers/ProjectData.ts";
+import { ProjectData } from "../../../types/ProjectData.ts";
 import { TagsBar } from "../../../components/tags-bar/TagsBar.tsx";
 import { BooleanTag } from "../../../components/tags-bar/tag/BooleanTag.tsx";
 import { shortDate } from "../../../helpers/PostData.ts";
@@ -45,14 +45,6 @@ const ProjectCardComponent = ({ project: props }: Props) => {
         )}
 
         <p className="project-description">{props.description}</p>
-
-        {/*{ props.images && props.images.length > 0 &&*/}
-        {/*  <div className="project-images">*/}
-        {/*    { props.images?.map((image) =>*/}
-        {/*      <img src={`${import.meta.env.BASE_URL}${image.src}`} alt={image.alt} />*/}
-        {/*    )}*/}
-        {/*  </div>*/}
-        {/*}*/}
 
         {props.images && props.images.length > 0 && (
           <div className="project-images">
