+++ src/helpers/PostData.ts
@@ -1,20 +1,3 @@
-export type PostData = {
-  title: string;
-  content: Content;
-  tags: string[];
-  date: string;
-};
-
-export enum ContentTypes {
-  DEFAULT = "default",
-}
-
-export type Content = {
-  type: ContentTypes.DEFAULT;
-  data: {
-    textContent?: string;
-  };
-};
 
 export const shortDate = (dateStr: string) => {
   const date = new Date(dateStr);
